// src/app/api/models/trade-cards-models-legal-entities-tabs-response-legal-entity-basic-info-response.ts
/* tslint:disable */

/**
 * Базовое описание юр. лица
 */
export interface TradeCardsModelsLegalEntitiesTabsResponseLegalEntityBasicInfoResponse {

  /**
   * Дата архивирования
   */
  archiveDate?: string;

  /**
   * Контактный Email
   */
  contactEmail?: string;

  /**
   * ФИО контактного лица
   */
  contactPersonName?: string;

  /**
   * Контактный телефон
   */
  contactPhone?: string;

  /**
   * Дата создания
   */
  createDate?: string;

  /**
   * Электронная почта
   */
  email?: string;

  /**
   * Идентификатор
   */
  id?: string;

  /**
   * Действующий / Архивный
   */
  isActive?: boolean;

  /**
   * Наименование
   */
  name?: string;

  /**
   * Адрес офиса
   */
  officeAddress?: string;

  /**
   * Пароль
   */
  password?: string;
}