// src/app/api/models/trade-cards-models-legal-entities-tabs-request-update-legal-entity-basic-info-request.ts
/* tslint:disable */

/**
 * Запрос на обновление базового описания юр. лица
 */
export interface TradeCardsModelsLegalEntitiesTabsRequestUpdateLegalEntityBasicInfoRequest {

  /**
   * Контактный Email
   */
  contactEmail?: string;

  /**
   * ФИО контактного лица
   */
  contactPersonName?: string;

  /**
   * Контактный телефон
   */
  contactPhone?: string;

  /**
   * Электронная почта
   */
  email?: string;

  /**
   * Наименование
   */
  name?: string;

  /**
   * Адрес офиса
   */
  officeAddress?: string;

  /**
   * Пароль
   */
  password?: string;
}